+++ src/Languages/fr.lua
@@ -490,6 +490,7 @@ LUI.src.Languages.fr.FR = {
     ["Unlock non-matches"] = "Déverrouiller les non-correspondances",
     ["Use"] = "Utiliser",
     ["Use LotRO move mode shortcut"] = "Utiliser le raccourci du mode déplacement de LotRO",
+    ["Close LUI windows with Esc"] = "Fermer les fenêtres LUI avec Échap",
     ["Use native LotRO UI scaling"] = "Utiliser la mise à l'échelle native de l'interface de LotRO",
     ["Vault"] = "Coffre",
     ["View"] = "Vue",